+++ src/main/scala/de/htwg/winesmeeper/Controller/ControllerTrait.scala
@@ -6,7 +6,7 @@ import javafx.scene.input.KeyCode
 
 import scala.util.Try
 
-trait ControllerTrait() extends Observable:
+trait ControllerTrait extends Observable:
   val undo: TurnCmdManagerTrait
   val sysCmd: SysCommandManagerTrait
   var gb: BoardTrait
@@ -21,11 +21,11 @@ trait ControllerTrait() extends Observable:
 
   def isSysCmd(cmd: String): Boolean
 
-  def doSysCmd(cmd: String, params: Vector[String] = Vector()): Option[String]
+  def doSysCmd(subID: Int, cmd: String, params: Vector[String]): Option[String]
 
-  def turn(cmd: String, x: Try[Int], y: Try[Int]): Try[Boolean]
+  def turn(subID: Int, cmd: String, x: Try[Int], y: Try[Int]): Try[Boolean]
 
-  def doShortCut(key: KeyCode): Option[String]
+  def doShortCut(observerID: Int, key: KeyCode): Option[String]
 
   def getSysCmdList: List[String]