+++ src/main/scala/de/htwg/winesmeeper/Controller/ImplController/Controller.scala
@@ -17,16 +17,16 @@ class Controller @Inject() (var gb: BoardTrait) extends ControllerTrait():
   override val undo: TurnCmdManagerTrait = undoMaker(this)
   override val sysCmd: SysCommandManagerTrait = injector.instance
   
-  override def turn(cmd: String, x: Try[Int], y: Try[Int]): Try[Boolean] = {
-    Try(state.turn(cmd.toLowerCase, x.get, y.get))
+  override def turn(observerID: Int, cmd: String, x: Try[Int], y: Try[Int]): Try[Boolean] = {
+    Try(state.turn(observerID, cmd.toLowerCase, x.get, y.get))
   }
 
   override def changeState(newState: String): Unit = state.changeState(newState)
 
   override def isSysCmd(cmd: String): Boolean = sysCmd.isSysCmd(cmd.toLowerCase())
   
-  override def doSysCmd(cmd: String, params: Vector[String] = Vector("no params")): Option[String] =
-    sysCmd.doSysCmd(this, cmd.toLowerCase(), params)
+  override def doSysCmd(observerID: Int, cmd: String, params: Vector[String]): Option[String] =
+    sysCmd.doSysCmd(observerID, this, cmd.toLowerCase(), params)
     
   override def getBoard: Vector[Vector[Int]] = gb.getBoard
   
@@ -38,7 +38,7 @@ class Controller @Inject() (var gb: BoardTrait) extends ControllerTrait():
   
   override def getSysCmdList: List[String] = sysCmd.getSysCmdList.map(sys => sys.cmd)
   
-  override def doShortCut(key: KeyCode): Option[String] = sysCmd.doShortCut(this, key)
+  override def doShortCut(observerID: Int, key: KeyCode): Option[String] = sysCmd.doShortCut(observerID, this, key)
 
   override def isVictory: Boolean = gb.isVictory
 
@@ -57,5 +57,5 @@ object Controller:
       Guice.createInjector(WinesmeeperModule).instance
     val undo = undoMaker(out)
     for fx <- xStart - 1 to xStart + 1; fy <- yStart - 1 to yStart + 1 do
-      if gb.in(fx, fy) then undo.doCmd("open", fx, fy)
+      if gb.in(fx, fy) then undo.doCmd(-1,"open", fx, fy)
     out