+++ src/main/scala/de/htwg/winesmeeper/Controller/ImplTurnCommands/FlagCommand.scala
@@ -8,7 +8,7 @@ import scala.util.{Success, Try}
 import com.google.inject.Guice
 import net.codingwell.scalaguice.InjectorExtensions._
 
-case class FlagCommand(ctrl: ControllerTrait, x: Int, y: Int) extends CommandTrait:
+case class FlagCommand(observerID_ : Int, ctrl: ControllerTrait, x: Int, y: Int) extends CommandTrait(observerID_):
   
   override def doStep(): Boolean =
     val f = ctrl.gb.getFieldAt(x, y)
@@ -23,7 +23,7 @@ case class FlagCommand(ctrl: ControllerTrait, x: Int, y: Int) extends CommandTra
 
   override def redoStep(): Boolean = doStep()
 
-  override def toString: String = f"flag($x, $y)"
+  override def toString: String = f"flag($observerID, $x, $y)"
   
 object FlagCOR extends CommandCORTrait:
   override val cmd = "flag"
@@ -34,6 +34,6 @@ object FlagCOR extends CommandCORTrait:
       |  mark this position as flag or remove the flag
       |""".stripMargin
 
-  override def buildCmd(cmd: String, x: Int, y: Int, ctrl: ControllerTrait): Try[CommandTrait] =
-    if cmd == this.cmd then Success(FlagCommand(ctrl, x, y)) else next.buildCmd(cmd,x,y,ctrl)
+  override def buildCmd(observerID: Int, cmd: String, x: Int, y: Int, ctrl: ControllerTrait): Try[CommandTrait] =
+    if cmd == this.cmd then Success(FlagCommand(observerID, ctrl, x, y)) else next.buildCmd(observerID, cmd,x,y,ctrl)