+++ src/main/scala/de/htwg/winesmeeper/Main.scala
@@ -1,9 +1,12 @@
 package de.htwg.winesmeeper
 
-import de.htwg.winesmeeper.Controller.ImplController.Controller
-import de.htwg.winesmeeper.Model.ImplBoard.Board
+import de.htwg.winesmeeper.Controller.ControllerTrait
+import de.htwg.winesmeeper.Model.BoardTrait
 import de.htwg.winesmeeper.aView.TUI.TUI
+import de.htwg.winesmeeper.WinesmeeperModule
 
+import com.google.inject.Guice
+import net.codingwell.scalaguice.InjectorExtensions._
 import scala.io.StdIn.readInt
 
 val initVals = new Array[Int](5)
@@ -12,8 +15,11 @@ val initVals = new Array[Int](5)
   for i <- 0 until 5 do
     println(getPrintString(i))
     initVals(i) = readInt
-  val gb = Board(initVals(0), initVals(1), initVals(2), initVals(3), initVals(4))
-  val ctrl = Controller(initVals(2), initVals(3), gb)
+  val injector = Guice.createInjector(WinesmeeperModule)
+  val bMaker: (Int, Int, Int, Int, Int) => BoardTrait = injector.instance
+  val ctrlMaker: (Int, Int, BoardTrait) => ControllerTrait = injector.instance
+  val gb = bMaker(initVals(0), initVals(1), initVals(2), initVals(3), initVals(4))
+  val ctrl = ctrlMaker(initVals(2), initVals(3), gb)
   new Thread(() => {
   val tui = TUI(ctrl)
   }).start()