+++ src/main/scala/de/htwg/winesmeeper/Model/BoardTrait.scala
@@ -13,6 +13,7 @@ trait BoardTrait:
   def getField(x: Int, y: Int): Int
 
   def getFieldAt (indX: Int, indY: Int): FieldTrait
+  
 
   def isVictory: Boolean