+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.3.0-dev.3](https://github.com/Atomarverseucht/minesweeper/compare/1.3.0-dev.2...1.3.0-dev.3) (2026-01-09)
+
+
+### Bug Fixes
+
+* generate doesn't generate ([0843150](https://github.com/Atomarverseucht/minesweeper/commit/08431507ad03b67a4b84fcd1018ec425c1616b03))
+
+# [1.3.0-dev.2](https://github.com/Atomarverseucht/minesweeper/compare/1.3.0-dev.1...1.3.0-dev.2) (2026-01-09)
+
+
+### Features
+
+* generate cmd to generate a new board ([3e38a5b](https://github.com/Atomarverseucht/minesweeper/commit/3e38a5b3b92ceb4c3fa1c2feaeb4ccf6c88f8213))
+
+# [1.3.0-dev.1](https://github.com/Atomarverseucht/minesweeper/compare/1.2.2...1.3.0-dev.1) (2026-01-08)
+
+
+### Features
+
+* Observer (Views) are now having IDs ([1655e30](https://github.com/Atomarverseucht/minesweeper/commit/1655e306a167ec2cb6e7248d602ae4abca7a95f4))
+
 ## [1.2.2](https://github.com/Atomarverseucht/minesweeper/compare/1.2.1...1.2.2) (2026-01-08)